backend/requirements.txt
aioredis==1.3.1
apturl==0.5.2
asgi-redis==1.4.3
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.7.0
cachetools==4.0.0
cbor==1.0.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==36.0.1
cupshelpers==1.0
Cython==0.29.14
daphne==3.0.2
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.3
django-channels==0.7.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
favicon==0.7.0
filelock==3.6.0
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
h11==0.13.0
hiredis==2.0.0
httplib2==0.14.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
install==1.3.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
msgpack==1.0.3
msgpack-python==0.5.6
netifaces==0.10.4
oauth2client==4.1.2
oauthlib==3.2.0
olefile==0.46
outcome==1.1.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.0
protobuf==3.6.1
psycopg2-binary==2.9.3
py-ubjson==0.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
PyDrive==1.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pypng==0.0.20
PyQRCode==1.2.1
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
python-snappy==0.5.3
PyTrie==0.2
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
redis==2.10.6
reportlab==3.5.34
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.0
SecretStorage==2.3.1
selenium==4.1.3
service-identity==21.1.0
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.3.12
sqlparse==0.4.2
systemd-python==234
trio==0.20.0
trio-websocket==0.9.2
Twisted==22.1.0
txaio==22.2.1
typing-extensions==4.1.1
u-msgpack-python==2.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.26.8
usb-creator==0.3.7
virtualenv==20.13.1
wadllib==1.3.3
wsaccel==0.6.2
wsproto==1.1.0
xkit==0.0.0
zope.interface==5.4.0
